Curl and firmware update compiles now
