+++ .vscode/tasks.json
@@ -15,6 +15,7 @@
 				"-I${workspaceFolder}/support/SDL2/include/SDL2",
 				"-I${workspaceFolder}/support/SDL2Image/include/SDL2",
 				"-I${workspaceFolder}/support",
+				"-I${workspaceFolder}/support/curl",
 				"-I${workspaceFolder}/src",
 				"-I${workspaceFolder}/src/libraries",
 				"-I${workspaceFolder}/src/engine",
@@ -96,8 +97,12 @@
 				"${workspaceFolder}/src/engine/SpikeAnalysis.h",
 				"${workspaceFolder}/src/engine/SpikeSorter.cpp",
 				"${workspaceFolder}/src/engine/SpikeSorter.h",
+				"${workspaceFolder}/src/engine/firmware/BSLFirmwareUpdater.cpp",
+				"${workspaceFolder}/src/engine/firmware/BSLFirmwareUpdater.h",
 				"${workspaceFolder}/src/engine/firmware/BYBFirmwareVO.cpp",
 				"${workspaceFolder}/src/engine/firmware/BYBFirmwareVO.h",
+				"${workspaceFolder}/src/engine/firmware/FirmwareUpdater.cpp",
+				"${workspaceFolder}/src/engine/firmware/FirmwareUpdater.h",
 				"${workspaceFolder}/src/libraries/tinyxml2.cpp",
 				"${workspaceFolder}/src/libraries/tinyxml2.h",
 				"${workspaceFolder}/src/native/PathsWin.cpp",
