+++ support/curl/curlrules.h
